Algo Trading/Algo Trade with Cursor/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
binance-futures-connector>=3.0.0 